TRUS-3178 Move All bar Subscription call from DesConnector to IfsConnector
