+++ app/connector/DesConnector.scala
@@ -20,16 +20,9 @@ import java.util.UUID
 
 import config.AppConfig
 import javax.inject.Inject
-import models._
-import models.existing_trust.{ExistingCheckRequest, ExistingCheckResponse}
-import models.get_trust.GetTrustResponse
-import models.registration.RegistrationResponse
 import models.tax_enrolments.SubscriptionIdResponse
-import models.variation.VariationResponse
 import play.api.Logging
 import play.api.http.HeaderNames
-import play.api.libs.json._
-import services.TrustsStoreService
 import uk.gov.hmrc.http.{HeaderCarrier, HttpClient}
 import utils.Constants._
 import utils.Session
@@ -37,24 +30,9 @@ import utils.Session
 import scala.concurrent.ExecutionContext.Implicits._
 import scala.concurrent.{ExecutionContext, Future}
 
-class DesConnector @Inject()(http: HttpClient, config: AppConfig, trustsStoreService: TrustsStoreService) extends Logging {
+class DesConnector @Inject()(http: HttpClient, config: AppConfig) extends Logging {
 
-  private lazy val trustsServiceUrl : String = s"${config.registerTrustsUrl}/trusts"
-  private lazy val matchTrustsEndpoint : String = s"$trustsServiceUrl/match"
-  private lazy val trustRegistrationEndpoint : String = s"$trustsServiceUrl/registration"
-  private lazy val getTrustOrEstateUrl: String =  s"${config.getTrustOrEstateUrl}/trusts"
-
-  def get4MLDTrustOrEstateEndpoint(utr: String): String = s"$getTrustOrEstateUrl/registration/$utr"
-
-  def get5MLDTrustOrEstateEndpoint(identifier: String): String = {
-    if (identifier.length == 10) {
-      s"$getTrustOrEstateUrl/registration/UTR/$identifier"
-    } else {
-      s"$getTrustOrEstateUrl/registration/URN/$identifier"
-    }
-  }
-
-  private lazy val trustVariationsEndpoint : String = s"${config.varyTrustOrEstateUrl}/trusts/variation"
+  private lazy val trustsServiceUrl : String = s"${config.desTrustsBaseUrl}/trusts"
 
   val ENVIRONMENT_HEADER = "Environment"
   val CORRELATION_HEADER = "CorrelationId"
@@ -67,38 +45,7 @@ class DesConnector @Inject()(http: HttpClient, config: AppConfig, trustsStoreSer
       CORRELATION_HEADER -> correlationId
     )
 
-  def checkExistingTrust(existingTrustCheckRequest: ExistingCheckRequest)
-                                 : Future[ExistingCheckResponse] = {
-
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.info(s"[Session ID: ${Session.id(hc)}] matching trust for correlationId: $correlationId")
-
-    val response = http.POST[JsValue, ExistingCheckResponse](matchTrustsEndpoint, Json.toJson(existingTrustCheckRequest))
-    (implicitly[Writes[JsValue]], ExistingCheckResponse.httpReads, implicitly[HeaderCarrier](hc),implicitly[ExecutionContext])
-
-    response
-  }
-
-  def registerTrust(registration: Registration)
-                            : Future[RegistrationResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.debug(s"[Session ID: ${Session.id(hc)}] registration payload ${Json.toJson(registration)}")
-
-    logger.info(s"[Session ID: ${Session.id(hc)}] registering trust for correlationId: $correlationId")
-
-    val response = http.POST[JsValue, RegistrationResponse](trustRegistrationEndpoint, Json.toJson(registration))
-    (implicitly[Writes[JsValue]], RegistrationResponse.httpReads, implicitly[HeaderCarrier](hc), implicitly[ExecutionContext])
-
-    response
-  }
-
-  def getSubscriptionId(trn: String): Future[SubscriptionIdResponse] = {
+   def getSubscriptionId(trn: String): Future[SubscriptionIdResponse] = {
 
     val correlationId = UUID.randomUUID().toString
 
@@ -114,38 +61,4 @@ class DesConnector @Inject()(http: HttpClient, config: AppConfig, trustsStoreSer
     response
   }
 
-  def getTrustInfo(identifier: String): Future[GetTrustResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc : HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.info(s"[Session ID: ${Session.id(hc)}][UTR/URN: $identifier]" +
-      s" getting playback for trust for correlationId: $correlationId")
-
-    trustsStoreService.is5mldEnabled.flatMap { is5MLD =>
-      if (is5MLD) {
-        http.GET[GetTrustResponse](get5MLDTrustOrEstateEndpoint(identifier))(GetTrustResponse.httpReads(identifier), implicitly[HeaderCarrier](hc), global)
-      } else {
-        http.GET[GetTrustResponse](get4MLDTrustOrEstateEndpoint(identifier))(GetTrustResponse.httpReads(identifier), implicitly[HeaderCarrier](hc), global)
-      }
-    }
-  }
-
-  def trustVariation(trustVariations: JsValue): Future[VariationResponse] = {
-    val correlationId = UUID.randomUUID().toString
-
-    implicit val hc: HeaderCarrier = HeaderCarrier(extraHeaders = desHeaders(correlationId))
-
-    logger.debug(s"[Session ID: ${Session.id(hc)}] variation payload $trustVariations}")
-
-    logger.info(s"[Session ID: ${Session.id(hc)}]" +
-      s" submitting trust variation for correlationId: $correlationId")
-
-    http.POST[JsValue, VariationResponse](trustVariationsEndpoint, Json.toJson(trustVariations))(
-      implicitly[Writes[JsValue]],
-      VariationResponse.httpReads,
-      implicitly[HeaderCarrier](hc),
-      implicitly[ExecutionContext]
-    )
-  }
 }
