+++ conf/application.conf
@@ -174,6 +174,14 @@ microservice {
             port=8500
         }
 
+        ifs-trusts {
+            protocol = http
+            host = localhost
+            port = 9802
+            token = devToken
+            environment = dev
+        }
+
         des-trusts {
             protocol = http
             host = localhost
