+++ it/uk/gov/hmrc/transformations/AddBusinessSettlorSpec.scala
@@ -26,7 +26,7 @@ import play.api.test.Helpers._
 import play.api.test.{FakeRequest, Helpers}
 import uk.gov.hmrc.auth.core.AffinityGroup.Organisation
 import uk.gov.hmrc.itbase.IntegrationTestBase
-import connector.DesConnector
+import connector.IfsConnector
 import controllers.actions.{FakeIdentifierAction, IdentifierAction}
 import models.get_trust.GetTrustSuccessResponse
 import play.api.Application
@@ -60,13 +60,13 @@ class AddBusinessSettlorSpec extends AsyncFreeSpec with MustMatchers with Mockit
     lazy val expectedGetAfterAddSettlorJson: JsValue =
       JsonUtils.getJsonValueFromFile("add-business-settlor-after-etmp-call.json")
 
-    val stubbedDesConnector = mock[DesConnector]
+    val stubbedDesConnector = mock[IfsConnector]
     when(stubbedDesConnector.getTrustInfo(any())).thenReturn(Future.successful(getTrustResponseFromDES))
 
     val application = applicationBuilder
       .overrides(
         bind[IdentifierAction].toInstance(new FakeIdentifierAction(Helpers.stubControllerComponents().parsers.default, Organisation)),
-        bind[DesConnector].toInstance(stubbedDesConnector)
+        bind[IfsConnector].toInstance(stubbedDesConnector)
       )
       .build()
