+++ it/uk/gov/hmrc/transformations/SetRelationSpec.scala
@@ -16,7 +16,7 @@
 
 package uk.gov.hmrc.transformations
 
-import connector.DesConnector
+import connector.IfsConnector
 import controllers.actions.{FakeIdentifierAction, IdentifierAction}
 import models.get_trust.GetTrustSuccessResponse
 import org.mockito.Matchers._
@@ -41,7 +41,7 @@ class SetRelationSpec extends AsyncFreeSpec with MustMatchers with MockitoSugar 
       val getTrustResponseFromDES : JsValue = JsonUtils
         .getJsonValueFromFile("trusts-etmp-received.json")
 
-      val stubbedDesConnector = mock[DesConnector]
+      val stubbedDesConnector = mock[IfsConnector]
 
       when(stubbedDesConnector.getTrustInfo(any()))
         .thenReturn(Future.successful(getTrustResponseFromDES.as[GetTrustSuccessResponse]))
@@ -49,7 +49,7 @@ class SetRelationSpec extends AsyncFreeSpec with MustMatchers with MockitoSugar 
       val application = applicationBuilder
         .overrides(
           bind[IdentifierAction].toInstance(new FakeIdentifierAction(Helpers.stubControllerComponents().parsers.default, Organisation)),
-          bind[DesConnector].toInstance(stubbedDesConnector)
+          bind[IfsConnector].toInstance(stubbedDesConnector)
         )
         .build()
